+++ src/components/LSVGOverlay.vue
@@ -1,0 +1,51 @@
+<script setup lang="ts">
+import { SVGOverlay } from 'leaflet'
+import { markRaw, nextTick, onMounted, ref, useAttrs } from 'vue'
+import { AddLayerInjection } from '../types/injectionKeys'
+import { assertInject, propsBinder, remapEvents } from '../utils.js'
+import {
+    setupSVGOverlay,
+    type SVGOverlayEmits,
+    type SVGOverlayProps,
+    svgOverlayPropsDefaults
+} from '../functions/svgOverlay.ts'
+
+const props = withDefaults(defineProps<SVGOverlayProps>(), svgOverlayPropsDefaults)
+const emit = defineEmits<SVGOverlayEmits>()
+
+const { ready, leafletObject } = useSVGOverlay()
+defineExpose({ ready, leafletObject })
+
+function useSVGOverlay() {
+    const leafletObject = ref<SVGOverlay>()
+    const ready = ref(false)
+
+    const addLayer = assertInject(AddLayerInjection)
+
+    const { options, methods } = setupSVGOverlay(props, leafletObject, emit)
+
+    onMounted(async () => {
+        leafletObject.value = markRaw<SVGOverlay>(
+            new SVGOverlay(props.svg, props.bounds, options)
+        )
+
+        const { listeners } = remapEvents(useAttrs())
+        leafletObject.value.on(listeners)
+        propsBinder(methods, leafletObject.value, props)
+        addLayer({
+            ...props,
+            ...methods,
+            leafletObject: leafletObject.value
+        })
+        ready.value = true
+        nextTick(() => emit('ready', leafletObject.value!))
+    })
+
+    return { ready, leafletObject }
+}
+</script>
+
+<template>
+    <div v-if="ready" style="display: none">
+    </div>
+</template>